Correct Non-GMT Time Zone based time calculation
This commit corrects few errors in the logic for time zone based time calculation.
For example, if the within time was given as "within "2017-08-** **:**:** +05:30",
the start time of the duration would map to 1501525800000 in milliseconds, This time is equal to July 31, 2017 6:30:00 PM
in GMT. In the previous implementation, the end time for this within duration was calculated, by taking this long value and
getting the next month, assuming it's GMT. Hence the next month is wrongly claculated as August and not September.
Furthermore, the system time zone is taken into consideration for event arrival time based executions.