Support list of values as the output of counting-pattern
When counting patterns are used such as `e1=StockStream<2:8>` and when
they are referred without indexes such as `e1.price` it collects the
price values from all the events in the counting pattern `e1` and
produce it as a list. Since list if not native to Siddhi the attribute
will be have `object` as its type.