Fix for https://github.com/siddhi-io/siddhi-map-json/issues/100
The trpProperties is set in the SourceMapper but it is removed by both SourceMapper and InputEventHandler.
For each event SourceMapper gets, it sets the trpProperties and hands over the event to InputEventHandler. After that removes the trpProperties.
For each event InputEventHandler gets, it removes the trpProperties too (it is unclear why it is doing this duplicate work).
Problem occurs when a batch of events comes to SourceMapper and it invokes the InputEventHandler multiple times (once for each event in the batch). Since InputEventHandler removes the trpProperties each time it gets an event; but the SourceMapper only sets it once for the whole batch, the 2nd event in the batch gets null for the trpProperties. This is the rootcause of this issue.

In this fix, we assume that only the SourceMapper submits events to the InputEventHandler. So setting/removal of trpProperties is handled solely by the SourceMapper.